chore: move function getExtraAccountsIds to helpers folders
